Move spiegelman benchmark file
